Add files via upload
Exercise 1.
Three resistances R1, R2, R3 are given. Calculate the value of the resistance R0 according to the formula: 1 / R0 = 1 / R1 + 1 / R2 + 1 / R3.
Test example: R1=2, R2=4, R3=8 R0 = 1.142857.
Exercise 2.
Given the length of a circle, calculate the area of a circle using a formula S=pi*R2, calculate the radius from the formula of a length of a circle: L = 2 * pi * R.
Note: pi = 3.14.
Exercise 3.
Calculate the traveled distance for the rectilinear uniformly accelerated motion using a formula S = v * t + (a * t2) / 2, where v - speed, t - time, and a - acceleration.

Topic: Variables and data types.
Exercise1.
The user enters a time from the keyboard in seconds. You must write a program which converts seconds that the user inputs, into hours, minutes, seconds and displays them on the screen.
Exercise2.
Write a program that converts a fractional number which is input from the keypad in a currency format. For example, a number 12.5 should be converted into the form 12grn 50 kopecks.
Exercise3.
Write a program that computes a speed of the runner that ran a distance. As while the program operates, the screen should look as is shown below:
Calculation of the running speed
Input a distance length (meters) = 1000. Input a time (minutes, seconds) = 3.25. Distance: 1000 m.
Time: 3 minutes 25 seconds = 205 seconds. You ran at a speed of 17.56 km / h.
Exercise4.
 The user specifies a price of one minute of outgoing call from one mobile operator to another, as well as the duration of the call is in minutes and seconds. It is
 necessary to calculate a cost of the call.
Exercise5.
Write a program that converts user-entered number of days into a number of complete weeks and remaining days. For example, the user has entered 17 days, the program should display 2 weeks and 3 days.